+++ lib/lightning/work_orders.ex
@@ -78,14 +78,15 @@ defmodule Lightning.WorkOrders do
       create_for(job, workflow: workflow, dataclip: dataclip, user: user)
   """
   @spec create_for(Trigger.t() | Job.t(), Multi.t(), [work_order_option()]) ::
-          {:ok, WorkOrder.t()} | {:error, Ecto.Changeset.t(WorkOrder.t())}
+          {:ok, WorkOrder.t()}
+          | {:error, Ecto.Changeset.t(WorkOrder.t()) | :workflow_deleted}
   def create_for(target, multi \\ Multi.new(), opts)
 
   def create_for(%Trigger{} = trigger, multi, opts) do
     multi
     |> Multi.put(:workflow, opts[:workflow])
     |> get_or_insert_dataclip(opts[:dataclip])
-    |> get_or_create_snapshot(opts[:workflow])
+    |> get_or_create_snapshot(opts[:workflow], opts[:actor])
     |> Multi.insert(:workorder, fn %{dataclip: dataclip, snapshot: snapshot} ->
       {without_run?, opts} = Keyword.pop(opts, :without_run, false)
 
@@ -110,7 +111,7 @@ defmodule Lightning.WorkOrders do
   def create_for(%Job{} = job, multi, opts) do
     multi
     |> Multi.put(:workflow, opts[:workflow])
-    |> get_or_create_snapshot()
+    |> get_or_create_snapshot(opts[:actor])
     |> Multi.insert(:workorder, build_for(job, opts |> Map.new()))
     |> Runs.enqueue()
     |> emit_and_return_work_order()
@@ -118,9 +119,16 @@ defmodule Lightning.WorkOrders do
 
   def create_for(%Manual{} = manual) do
     Multi.new()
+    |> Multi.run(:workflow_deleted?, fn _repo, _changes ->
+      if manual.workflow.deleted_at do
+        {:error, :workflow_deleted}
+      else
+        {:ok, false}
+      end
+    end)
     |> get_or_insert_dataclip(manual)
     |> Multi.put(:workflow, manual.workflow)
-    |> get_or_create_snapshot()
+    |> get_or_create_snapshot(manual.created_by)
     |> Multi.insert(:workorder, fn %{dataclip: dataclip, snapshot: snapshot} ->
       build_for(manual.job, %{
         workflow: manual.workflow,
@@ -146,7 +154,7 @@ defmodule Lightning.WorkOrders do
     {:error, changeset}
   end
 
-  defp get_or_create_snapshot(multi, workflow \\ nil, name \\ :snapshot) do
+  defp get_or_create_snapshot(multi, workflow \\ nil, name \\ :snapshot, actor) do
     multi
     |> Multi.merge(fn
       %{^name => _snapshot} ->
@@ -155,7 +163,7 @@ defmodule Lightning.WorkOrders do
 
       changes ->
         workflow = workflow || changes[:workflow]
-        Snapshot.get_or_create_latest_for(Multi.new(), name, workflow)
+        Snapshot.get_or_create_latest_for(Multi.new(), name, workflow, actor)
     end)
   end
 
@@ -194,7 +202,7 @@ defmodule Lightning.WorkOrders do
     if snapshot = attrs |> Map.get(:snapshot) do
       changeset |> put_assoc(:snapshot, snapshot)
     else
-      Snapshot.get_or_create_latest_for(attrs[:workflow])
+      Snapshot.get_or_create_latest_for(attrs[:workflow], attrs[:actor])
       |> case do
         {:ok, snapshot} ->
           changeset |> put_assoc(:snapshot, snapshot)
@@ -283,7 +291,7 @@ defmodule Lightning.WorkOrders do
           Step.t() | Ecto.UUID.t(),
           [work_order_option(), ...]
         ) ::
-          {:ok, Run.t()} | {:error, Ecto.Changeset.t()}
+          {:ok, Run.t()} | {:error, Ecto.Changeset.t() | :workflow_deleted}
   def retry(%Run{id: run_id}, %Step{id: step_id}, opts) do
     retry(run_id, step_id, opts)
   end
@@ -310,6 +318,13 @@ defmodule Lightning.WorkOrders do
         preload: [:job]
       )
     end)
+    |> Multi.run(:workflow_deleted?, fn _repo, %{run: run} ->
+      if run.work_order.workflow.deleted_at do
+        {:error, :workflow_deleted}
+      else
+        {:ok, false}
+      end
+    end)
     |> Multi.run(:input_dataclip_id, fn
       _repo, %{step: %Step{input_dataclip_id: input_dataclip_id}} ->
         {:ok, input_dataclip_id}
@@ -332,7 +347,7 @@ defmodule Lightning.WorkOrders do
     |> Multi.run(:workflow, fn _repo, %{run: run} ->
       {:ok, run.work_order.workflow}
     end)
-    |> get_or_create_snapshot()
+    |> get_or_create_snapshot(creating_user)
     |> Multi.insert(:new_run, fn %{
                                    run: run,
                                    step: step,
@@ -501,7 +516,10 @@ defmodule Lightning.WorkOrders do
           retry(run_step.run_id, run_step.step_id, opts)
         end)
 
-      {:ok, Enum.count(results, fn result -> match?({:ok, _}, result) end), 0}
+      success_count =
+        Enum.count(results, fn result -> match?({:ok, _}, result) end)
+
+      {:ok, success_count, Enum.count(results) - success_count}
     end
   end
 
@@ -525,7 +543,7 @@ defmodule Lightning.WorkOrders do
       snapshot_op = "snapshot-#{workflow.id}"
 
       multi
-      |> get_or_create_snapshot(workflow, snapshot_op)
+      |> get_or_create_snapshot(workflow, snapshot_op, creating_user)
       |> Multi.insert(run_op, fn %{^snapshot_op => snapshot} ->
         starting_job = determine_starting_job(workorder)
 
@@ -672,6 +690,8 @@ defmodule Lightning.WorkOrders do
   defp fetch_retriable_workorders(workorder_ids) do
     workorder_ids
     |> workorders_with_dataclips_query()
+    |> join(:inner, [wo], wf in assoc(wo, :workflow), as: :workflow)
+    |> where([workflow: wf], is_nil(wf.deleted_at))
     |> Repo.all()
   end