+++ test/lightning/invocation_test.exs
@@ -11,6 +11,7 @@ defmodule Lightning.InvocationTest do
   require SearchParams
 
   defp build_workflow(opts) do
+    actor = insert(:user)
     job = build(:job)
     trigger = build(:trigger)
 
@@ -22,7 +23,7 @@ defmodule Lightning.InvocationTest do
       |> insert()
 
     {:ok, snapshot} =
-      Lightning.Workflows.Snapshot.get_or_create_latest_for(workflow)
+      Lightning.Workflows.Snapshot.get_or_create_latest_for(workflow, actor)
 
     %{
       workflow: workflow,