+++ test/lightning/work_orders_test.exs
@@ -4,10 +4,12 @@ defmodule Lightning.WorkOrdersTest do
   import Lightning.Factories
 
   alias Ecto.Multi
+  alias Lightning.Auditing.Audit
   alias Lightning.Extensions.MockUsageLimiter
   alias Lightning.Extensions.UsageLimiting.Action
   alias Lightning.Extensions.Message
   alias Lightning.KafkaTriggers.TriggerKafkaMessageRecord
+  alias Lightning.Workflows.Snapshot
   alias Lightning.WorkOrders
   alias Lightning.WorkOrders.Events
   alias Lightning.WorkOrders.RetryManyWorkOrdersJob
@@ -182,6 +184,29 @@ defmodule Lightning.WorkOrdersTest do
              |> Repo.get_by(trigger_id: trigger.id) != nil
     end
 
+    @tag trigger_type: :cron
+    test "with a trigger assigns the provided actor to the event", %{
+      snapshot: snapshot,
+      trigger: trigger,
+      workflow: workflow
+    } do
+      Repo.delete!(snapshot)
+
+      Lightning.WorkOrders.subscribe(workflow.project_id)
+
+      dataclip = insert(:dataclip)
+      %{id: actor_id} = actor = insert(:user)
+
+      WorkOrders.create_for(
+        trigger,
+        actor: actor,
+        dataclip: dataclip,
+        workflow: workflow
+      )
+
+      assert %{actor_id: ^actor_id} = Repo.one(Audit)
+    end
+
     test "with a manual workorder", context do
       %{workflow: workflow, job: job, snapshot: snapshot} = context
       user = insert(:user)
@@ -230,6 +255,38 @@ defmodule Lightning.WorkOrdersTest do
       assert_received %Events.WorkOrderCreated{
         work_order: %{id: ^workorder_id}
       }
+    end
+
+    test "assigns the created_by of the Manual as the actor for the audit", %{
+      job: job,
+      snapshot: snapshot,
+      workflow: workflow
+    } do
+      Repo.delete(snapshot)
+
+      %{id: user_id} = user = insert(:user)
+      project_id = workflow.project_id
+      Lightning.WorkOrders.subscribe(project_id)
+
+      {:ok, manual} =
+        Lightning.WorkOrders.Manual.new(
+          %{
+            "body" =>
+              Jason.encode!(%{
+                "key_left" => "value_left",
+                "configuration" => %{"password" => "secret"}
+              })
+          },
+          workflow: workflow,
+          project: workflow.project,
+          job: job,
+          created_by: user
+        )
+        |> Ecto.Changeset.apply_action(:validate)
+
+      WorkOrders.create_for(manual)
+
+      assert %{actor_id: ^user_id} = Repo.one!(Audit)
     end
 
     test "with a job", %{job: job, workflow: workflow, snapshot: snapshot} do
@@ -288,6 +345,34 @@ defmodule Lightning.WorkOrdersTest do
 
       assert TriggerKafkaMessageRecord
              |> Repo.get_by(trigger_id: trigger.id) != nil
+    end
+
+    test "with a job, assigns the actor to the audit if snapshot created", %{
+      job: job,
+      workflow: workflow
+    } do
+      Repo.delete_all(Snapshot)
+
+      project_id = workflow.project_id
+
+      project =
+        Repo.get(Lightning.Projects.Project, project_id)
+        |> Lightning.Projects.Project.changeset(%{retention_policy: :erase_all})
+        |> Repo.update!()
+
+      dataclip = insert(:dataclip, project: project)
+      user = insert(:user)
+      %{id: actor_id} = actor = insert(:user)
+
+      WorkOrders.create_for(
+        job,
+        actor: actor,
+        dataclip: dataclip,
+        workflow: workflow,
+        created_by: user
+      )
+
+      assert %{actor_id: ^actor_id} = Repo.one!(Audit)
     end
   end
 
@@ -349,7 +434,7 @@ defmodule Lightning.WorkOrdersTest do
 
       # This isn't the best place to test for this specific case.
       Lightning.Workflows.change_workflow(workflow, %{name: "new name"})
-      |> Lightning.Workflows.save_workflow()
+      |> Lightning.Workflows.save_workflow(insert(:user))
 
       snapshot2 = Lightning.Workflows.Snapshot.get_current_for(workflow)
 
@@ -572,6 +657,87 @@ defmodule Lightning.WorkOrdersTest do
       assert_received %Events.WorkOrderUpdated{
         work_order: %{id: ^workorder_id}
       }
+    end
+
+    test "returns error in case the workflow is deleted", %{
+      workflow: workflow,
+      snapshot: snapshot,
+      trigger: trigger,
+      jobs: [job | _rest]
+    } do
+      user = insert(:user)
+      dataclip = insert(:dataclip)
+      # create existing complete run
+      %{runs: [run]} =
+        insert(:workorder,
+          workflow: workflow,
+          snapshot: snapshot,
+          trigger: trigger,
+          dataclip: dataclip,
+          runs: [
+            %{
+              state: :failed,
+              dataclip: dataclip,
+              snapshot: snapshot,
+              starting_trigger: trigger,
+              steps: [
+                step = insert(:step, job: job, input_dataclip: dataclip)
+              ]
+            }
+          ]
+        )
+
+      # delete workflow
+      workflow
+      |> Ecto.Changeset.change(%{
+        deleted_at: DateTime.utc_now() |> DateTime.truncate(:second)
+      })
+      |> Repo.update!()
+
+      assert {:error, :workflow_deleted} =
+               WorkOrders.retry(run, step, created_by: user)
+    end
+
+    test "if snapshot is created, uses creating user as audit actor", %{
+      workflow: workflow,
+      snapshot: snapshot,
+      trigger: trigger,
+      jobs: [job_a, job_b, job_c]
+    } do
+      %{id: user_id} = user = insert(:user)
+      dataclip = insert(:dataclip)
+      output_dataclip = insert(:dataclip)
+
+      # create existing complete run
+      %{runs: [run]} =
+        insert(:workorder,
+          workflow: workflow,
+          trigger: trigger,
+          dataclip: dataclip,
+          runs: [
+            %{
+              state: :failed,
+              dataclip: dataclip,
+              starting_trigger: trigger,
+              steps: [
+                insert(:step,
+                  job: job_a,
+                  input_dataclip: dataclip,
+                  output_dataclip: output_dataclip
+                ),
+                second_step =
+                  insert(:step, job: job_b, input_dataclip: output_dataclip),
+                insert(:step, job: job_c)
+              ]
+            }
+          ]
+        )
+
+      Repo.delete(snapshot)
+
+      WorkOrders.retry(run, second_step, created_by: user)
+
+      assert %{actor_id: ^user_id} = Repo.one!(Audit)
     end
   end
 
@@ -2036,6 +2202,98 @@ defmodule Lightning.WorkOrdersTest do
                )
       end)
     end
+
+    test "retrying multiple workorders only retries workorders with workflows that havent been deleted",
+         %{
+           jobs: [job_a | _rest],
+           snapshot: snapshot,
+           trigger: trigger,
+           user: user,
+           workflow: workflow
+         } do
+      workorder_1 =
+        insert(:workorder,
+          workflow: workflow,
+          trigger: trigger,
+          dataclip: build(:dataclip),
+          snapshot: snapshot,
+          runs: [
+            %{
+              state: :failed,
+              dataclip: build(:dataclip),
+              starting_trigger: trigger,
+              snapshot: snapshot,
+              steps: [
+                insert(:step,
+                  job: job_a,
+                  input_dataclip: build(:dataclip),
+                  output_dataclip: build(:dataclip)
+                )
+              ]
+            }
+          ]
+        )
+
+      deleted_workflow =
+        insert(:simple_workflow,
+          project: workflow.project,
+          deleted_at: DateTime.utc_now()
+        )
+
+      deleted_workflow_job = hd(deleted_workflow.jobs)
+
+      snapshot_2 =
+        Lightning.Workflows.Snapshot.build(deleted_workflow) |> Repo.insert!()
+
+      workorder_2 =
+        insert(:workorder,
+          workflow: deleted_workflow,
+          trigger: hd(deleted_workflow.triggers),
+          dataclip: build(:dataclip),
+          snapshot: snapshot_2,
+          runs: [
+            %{
+              state: :failed,
+              dataclip: build(:dataclip),
+              starting_trigger: hd(deleted_workflow.triggers),
+              snapshot: snapshot_2,
+              steps: [
+                insert(:step,
+                  job: deleted_workflow_job,
+                  input_dataclip: build(:dataclip),
+                  output_dataclip: build(:dataclip)
+                )
+              ]
+            }
+          ]
+        )
+
+      refute Repo.get_by(Lightning.Run,
+               work_order_id: workorder_1.id,
+               starting_job_id: job_a.id
+             )
+
+      refute Repo.get_by(Lightning.Run,
+               work_order_id: workorder_2.id,
+               starting_job_id: deleted_workflow_job.id
+             )
+
+      {:ok, 1, 1} =
+        WorkOrders.retry_many([workorder_2, workorder_1],
+          created_by: user,
+          project_id: workflow.project_id
+        )
+
+      assert Repo.get_by(Lightning.Run,
+               work_order_id: workorder_1.id,
+               starting_job_id: job_a.id
+             )
+
+      refute Repo.get_by(Lightning.Run,
+               work_order_id: workorder_2.id,
+               starting_job_id: deleted_workflow_job.id
+             )
+    end
   end
 
   describe "retry_many/2 for RunSteps" do
@@ -2406,7 +2664,7 @@ defmodule Lightning.WorkOrdersTest do
                starting_job_id: job_a.id
              )
 
-      {:ok, 1, 0} =
+      {:ok, 1, 1} =
         WorkOrders.retry_many([run_step_2_a, run_step_1_a],
           created_by: user,
           project_id: workflow.project_id
@@ -2444,6 +2702,50 @@ defmodule Lightning.WorkOrdersTest do
       {:ok, work_order} = WorkOrders.update_state(run)
 
       assert work_order.state == :running
+    end
+  end
+
+  describe ".enqueue_many_for_retry/2" do
+    test "assigns the creating user as the auditing actor" do
+      dataclip = insert(:dataclip)
+      workflow = insert(:simple_workflow)
+
+      %{id: work_order_id} =
+        insert(:workorder, workflow: workflow, dataclip: dataclip)
+
+      %{id: user_id} = insert(:user)
+
+      WorkOrders.enqueue_many_for_retry([work_order_id], user_id)
+
+      assert %{actor_id: ^user_id} = Repo.one(Audit)
+    end
+  end
+
+  describe ".build" do
+    test "uses the provided actor for the snapshot audit event" do
+      dataclip = insert(:dataclip)
+      workflow = insert(:simple_workflow)
+      %{id: actor_id} = actor = insert(:user)
+
+      WorkOrders.build(%{dataclip: dataclip, workflow: workflow, actor: actor})
+
+      assert %{actor_id: ^actor_id} = Repo.one(Audit)
+    end
+  end
+
+  describe ".build_for" do
+    test "uses the provided actor for the snapshot audit event" do
+      dataclip = insert(:dataclip)
+      trigger = insert(:trigger)
+      workflow = insert(:simple_workflow)
+      %{id: actor_id} = actor = insert(:user)
+
+      WorkOrders.build_for(
+        trigger,
+        %{dataclip: dataclip, workflow: workflow, actor: actor}
+      )
+
+      assert %{actor_id: ^actor_id} = Repo.one(Audit)
     end
   end
 end