+++ test/lightning_web/controllers/webhooks_controller_test.exs
@@ -8,6 +8,7 @@ defmodule LightningWeb.WebhooksControllerTest do
   alias Lightning.Extensions.MockUsageLimiter
   alias Lightning.Extensions.StubUsageLimiter
 
+  alias Lightning.Auditing.Audit
   alias Lightning.Repo
   alias Lightning.Runs
   alias Lightning.WorkOrders
@@ -178,6 +179,16 @@ defmodule LightningWeb.WebhooksControllerTest do
 
       assert Runs.get_dataclip_request(run) ==
                ~s({\"path\": [\"i\", \"#{trigger_id}\"], \"method\": \"POST\", \"headers\": {\"content-type\": \"multipart/mixed; boundary=plug_conn_test\"}, \"query_params\": {\"moar\": \"things\", \"extra\": \"stuff\"}})
+    end
+
+    test "assigns the trigger as the actor for the audit event", %{conn: conn} do
+      %{triggers: [%{id: trigger_id}]} =
+        insert(:simple_workflow) |> Lightning.Repo.preload(:triggers)
+
+      message = %{"foo" => "bar"}
+      post(conn, "/i/#{trigger_id}", message)
+
+      assert %{actor_id: ^trigger_id} = Repo.one!(Audit)
     end
 
     test "returns 415 when client sends xml", %{conn: conn} do