+++ test/support/factories.ex
@@ -581,7 +581,10 @@ defmodule Lightning.Factories do
   end
 
   def work_order_for(trigger_or_job, attrs) do
-    Lightning.WorkOrders.build_for(trigger_or_job, Map.new(attrs))
+    Lightning.WorkOrders.build_for(
+      trigger_or_job,
+      Map.new(attrs) |> Map.merge(%{actor: insert(:user)})
+    )
     |> Ecto.Changeset.apply_changes()
   end